+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.20.0 (DEV)
 
 - Add: Experimental support for multiple languages. Russian language added.
+- Add: Display demo duration in Demo/Signature edit dialog.
 - Fix: FireFox cannot read text from clipboard (open/import test url).
 - Fix: Open/import level by test/demo url loss local options.
 - Fix: Export test/demo url could miss `pd=` (spec ports db) param, when spec